Move hash functions to security module
