+++ datameta/security.py
@@ -22,8 +22,19 @@ from pyramid.httpexceptions import HTTPFound, HTTPUnauthorized
 
 from .models import User, ApiKey
 
+import bcrypt
 import logging
 log = logging.getLogger(__name__)
+
+def hash_password(pw):
+    """Hash a password and return the salted hash"""
+    pwhash = bcrypt.hashpw(pw.encode('utf8'), bcrypt.gensalt())
+    return pwhash.decode('utf8')
+
+def check_password(pw, hashed_pw):
+    """Check a password against a salted hash"""
+    expected_hash = hashed_pw.encode('utf8')
+    return bcrypt.checkpw(pw.encode('utf8'), expected_hash)
 
 def get_bearer_token(request):
     """Extracts a Bearer authentication token from the request and returns it if present, None
